requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.1
celery==5.3.6
certifi==2025.8.3
cfgv==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.5
distlib==0.4.0
Django==4.2.1
django-celery-beat==2.8.1
django-timezone-field==7.1
filelock==3.19.1
flower==2.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
identify==2.6.14
idna==3.10
kombu==5.5.4
nodeenv==1.9.1
packaging==25.0
platformdirs==4.4.0
pre_commit==4.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
whitenoise==6.9.0
